# gh-117648: Improve performance of os.join by replacing map with a direct method …