--- a/Lib/posixpath.py
+++ b/Lib/posixpath.py
@@ -79,7 +79,8 @@ def join(a, *p):
     try:
         if not p:
             path[:0] + sep  #23780: Ensure compatible data type even if p is null.
-        for b in map(os.fspath, p):
+        for b in p:
+            b = os.fspath(b)
             if b.startswith(sep):
                 path = b
             elif not path or path.endswith(sep):