--- a/Misc/NEWS.d/next/Core and Builtins/2024-04-08-20-26-15.gh-issue-117648.NzVEa7.rst
+++ b/Misc/NEWS.d/next/Core and Builtins/2024-04-08-20-26-15.gh-issue-117648.NzVEa7.rst
@@ -1,0 +1,1 @@
+Speedup :func:`os.path.join` by up to 6% on Windows and up to ?% on Unix.