…call

Benchmark:

- `ntpath.py`
    ```bat
    ::test.bat
    @echo off
    echo 1 item && python -m timeit -s "import before.ntpath" "before.ntpath.join('foo')" && python -m timeit -s "import after.ntpath" "after.ntpath.join('foo')"
    echo 10 items && python -m timeit -s "import before.ntpath; paths = ['foo'] * 10" "before.ntpath.join(*paths)" && python -m timeit -s "import after.ntpath; paths = ['foo'] * 10" "after.ntpath.join(*paths)"
    echo 100 items && python -m timeit -s "import before.ntpath; paths = ['foo'] * 100" "before.ntpath.join(*paths)" && python -m timeit -s "import after.ntpath; paths = ['foo'] * 100" "after.ntpath.join(*paths)"
    ```

    ```none
    1 item
    500000 loops, best of 5: 699 nsec per loop # before
    500000 loops, best of 5: 612 nsec per loop # after
    # -> 1.14x faster
    10 items
    50000 loops, best of 5: 5.47 usec per loop # before
    50000 loops, best of 5: 5.18 usec per loop # after
    # -> 1.06x faster
    100 items
    5000 loops, best of 5: 54.6 usec per loop # before
    5000 loops, best of 5: 51.9 usec per loop # after
    # -> 1.05x faster
    ```
- `posixpath.py`
    ```bat
    # test.sh
    echo 1 item && python -m timeit -s "import before.posixpath" "before.posixpath.join('foo')" && python -m timeit -s "import after.posixpath" "after.posixpath.join('foo')"
    echo 10 items && python -m timeit -s "import before.posixpath; paths = ['foo'] * 10" "before.posixpath.join(*paths)" && python -m timeit -s "import after.posixpath; paths = ['foo'] * 10" "after.posixpath.join(*paths)"
    echo 100 items && python -m timeit -s "import before.posixpath; paths = ['foo'] * 100" "before.posixpath.join(*paths)" && python -m timeit -s "import after.posixpath; paths = ['foo'] * 100" "after.posixpath.join(*paths)"
    ```
    
    ```none
    1 item
    1000000 loops, best of 5: 325 nsec per loop # before
    1000000 loops, best of 5: 241 nsec per loop # after
    # -> 1.35x faster
    10 items
    100000 loops, best of 5: 3.33 usec per loop # before
    100000 loops, best of 5: 3.32 usec per loop # after
    # -> no difference
    100 items
    10000 loops, best of 5: 33.8 usec per loop # before
    10000 loops, best of 5: 34.1 usec per loop # after
    # -> no difference
    ```

<!-- gh-issue-number: gh-117648 -->
* Issue: gh-117648
<!-- /gh-issue-number -->